@@ -6,12 +6,13 @@
 
     <groupId>com.fireflyest</groupId>
     <artifactId>GlobalMarket</artifactId>
-    <version>3.0.9-SNAPSHOT</version>
+    <version>3.1.1</version>
+
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>16</maven.compiler.source>
-        <maven.compiler.target>16</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
     <build>
@@ -19,12 +20,16 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
-                <version>3.2.4</version>
+                <version>3.3.0</version>
                 <configuration>
                     <relocations>
                         <relocation>
                             <pattern>com.cryptomorin.xseries</pattern>
-                            <shadedPattern>com.fireflyest.gui.util</shadedPattern>
+                            <shadedPattern>com.fireflyest.gui</shadedPattern>
+                        </relocation>
+                        <relocation>
+                            <pattern>org.bstats</pattern>
+                            <shadedPattern>com.fireflyest.gui</shadedPattern>
                         </relocation>
                     </relocations>
                     <artifactSet>
@@ -43,6 +48,8 @@
                         </excludes>
                         <includes>
                             <include>com.github.cryptomorin:XSeries</include>
+                            <include>org.bstats:bstats-bukkit</include>
+                            <include>org.bstats:bstats-base</include>
                         </includes>
                     </artifactSet>
                 </configuration>
@@ -83,6 +90,12 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.bstats</groupId>
+            <artifactId>bstats-bukkit</artifactId>
+            <version>3.0.0</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
             <version>1.16.2-R0.1-SNAPSHOT</version>
@@ -91,12 +104,12 @@
         <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
-            <version>8.0.20</version>
+            <version>8.0.22</version>
         </dependency>
         <dependency>
             <groupId>org.xerial</groupId>
             <artifactId>sqlite-jdbc</artifactId>
-            <version>3.23.1</version>
+            <version>3.36.0.3</version>
         </dependency>
         <dependency>
             <groupId>com.github.MilkBowl</groupId>
@@ -107,12 +120,12 @@
         <dependency>
             <groupId>com.github.cryptomorin</groupId>
             <artifactId>XSeries</artifactId>
-            <version>7.2.1</version>
+            <version>8.8.0</version>
         </dependency>
         <dependency>
             <groupId>org.jetbrains</groupId>
             <artifactId>annotations</artifactId>
-            <version>18.0.0</version>
+            <version>23.0.0</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
