@@ -3,6 +3,7 @@ package com.fireflyest.market.listener;
 import com.cryptomorin.xseries.XMaterial;
 import com.cryptomorin.xseries.XSound;
 import com.fireflyest.gui.api.ViewGuide;
+import com.fireflyest.gui.api.ViewPage;
 import com.fireflyest.gui.event.ViewClickEvent;
 import com.fireflyest.market.GlobalMarket;
 import com.fireflyest.market.bean.User;
@@ -14,6 +15,7 @@ import com.fireflyest.market.util.ChatUtils;
 import com.fireflyest.market.util.ConvertUtils;
 import com.fireflyest.market.util.ItemUtils;
 import com.fireflyest.market.util.TimeUtils;
+import com.fireflyest.market.view.SellPage;
 import org.bukkit.Sound;
 import org.bukkit.block.Block;
 import org.bukkit.block.Sign;
@@ -102,6 +104,48 @@ public class PlayerEventListener implements Listener {
             return;
         }
 
+        // 快捷上架
+        ViewPage page = guide.getUsingPage(player.getName());
+        if (page instanceof SellPage sellPage){
+            switch (value) {
+                case "add1" -> {
+                    sellPage.addPrice(1);
+                    return;
+                }
+                case "add10" -> {
+                    sellPage.addPrice(10);
+                    return;
+                }
+                case "add100" -> {
+                    sellPage.addPrice(100);
+                    return;
+                }
+                case "reduce1" -> {
+                    sellPage.reducePrice(1);
+                    return;
+                }
+                case "reduce10" -> {
+                    sellPage.reducePrice(10);
+                    return;
+                }
+                case "reduce100" -> {
+                    sellPage.reducePrice(100);
+                    return;
+                }
+                case "amount" -> {
+                    if (event.isLeftClick()) {
+                        sellPage.raise(1);
+                    } else if (event.isRightClick()) {
+                        sellPage.lessen(1);
+                    }
+                    return;
+                }
+                default -> {
+                }
+            }
+        }
+
+        // 其他按钮点击，执行指令
         if (event.isShiftClick()){ // 下架
             if (value.contains("affair")){
                 marketAffair.affairCancel(player, ConvertUtils.parseInt(value.split(" ")[1]));
@@ -122,10 +166,10 @@ public class PlayerEventListener implements Listener {
         // 打开交易记录
         if(event.hasItem()){
             ItemStack item = event.getItem();
-            if(item == null)return;
-            if(!item.getType().equals(XMaterial.WRITTEN_BOOK.parseMaterial()))return;
-            String value = ItemUtils.getItemValue(item);
-            if(value.equals("record"))item.setAmount(0);
+            if(item != null && item.getType().equals(XMaterial.WRITTEN_BOOK.parseMaterial())) {
+                String value = ItemUtils.getItemValue(item);
+                if(value.equals("record"))item.setAmount(0);
+            }
         }
 
         // 右键牌子
@@ -138,7 +182,7 @@ public class PlayerEventListener implements Listener {
             if(sign.getLine(0).contains("GlobeMarket")){
                 Player player = event.getPlayer();
 
-                if(player.isSneaking()){
+                if("quick".equals(sign.getLine(1))){
                     player.performCommand("market quick");
                 }else {
                     player.performCommand("market");
