@@ -1,9 +1,9 @@
 package com.fireflyest.market.util;
 
-import com.cryptomorin.xseries.XMaterial;
 import org.bukkit.Material;
 
 import java.util.HashMap;
+import java.util.Locale;
 import java.util.Map;
 
 public class TranslateUtils {
@@ -16,11 +16,11 @@ public class TranslateUtils {
         if(itemList.containsKey(type)){
             return TranslateUtils.itemList.get(type);
         }else {
-            return XMaterial.toWord(material);
+            return type.replace("_", " ").toLowerCase(Locale.ROOT);
         }
     }
 
-    private static final Map<String, String> itemList = new HashMap<String, String>(){
+    private static final Map<String, String> itemList = new HashMap<>() {
         {
             put("STONE", "石头");
             put("GRANITE", "花岗岩");
@@ -606,6 +606,770 @@ public class TranslateUtils {
             put("WRITTEN_BOOK", "成书");
             put("REDSTONE", "红石");
             put("DIODE", "红石中继器");
+
+            // *
+            put("ACACIA_BOAT", "金合欢船");
+            put("ACACIA_BUTTON", "金合欢按钮");
+            put("ACACIA_CHEST_BOAT", "金合欢胸部船");
+            put("ACACIA_DOOR", "金合欢门");
+            put("ACACIA_FENCE", "金合欢栅栏");
+            put("ACACIA_FENCE_GATE", "金合欢篱笆门");
+            put("ACACIA_PLANKS", "金合欢木板");
+            put("ACACIA_PRESSURE_PLATE", "金合欢压力板");
+            put("ACACIA_SAPLING", "金合欢树苗");
+            put("ACACIA_SIGN", "金合欢牌子");
+            put("ACACIA_SLAB", "金合欢木台阶");
+            put("ACACIA_TRAPDOOR", "金合欢活板门");
+            put("ACACIA_WALL_SIGN", "金合欢墙牌");
+            put("ACACIA_WOOD", "金合欢木");
+            put("ACTIVATOR_RAIL", "激活铁轨");
+            put("AIR", "空气");
+            put("AMETHYST_BLOCK", "紫晶块");
+            put("AMETHYST_CLUSTER", "紫晶簇");
+            put("AMETHYST_SHARD", "紫水晶碎片");
+            put("ANCIENT_DEBRIS", "古老的碎片");
+            put("ANDESITE_SLAB", "安山岩台阶");
+            put("ANDESITE_STAIRS", "安山岩楼梯");
+            put("ANDESITE_WALL", "安山岩墙");
+            put("ANVIL", "铁砧");
+            put("AXOLOTL_BUCKET", "蝾螈桶");
+            put("AXOLOTL_SPAWN_EGG", "蝾螈刷怪蛋");
+            put("AZALEA", "杜鹃花");
+            put("AZALEA_LEAVES", "杜鹃花的叶子");
+            put("BAKED_POTATO", "烤土豆");
+            put("BAMBOO", "竹子");
+            put("BAMBOO_SAPLING", "竹树苗");
+            put("BARREL", "桶");
+            put("BARRIER", "屏障");
+            put("BASALT", "玄武岩");
+            put("BAT_SPAWN_EGG", "蝙蝠刷怪蛋");
+            put("BEACON", "信标");
+            put("BEDROCK", "基岩");
+            put("BEE_NEST", "蜜蜂窝");
+            put("BEE_SPAWN_EGG", "蜜蜂刷怪蛋");
+            put("BEEF", "牛肉");
+            put("BEEHIVE", "蜂巢");
+            put("BEETROOT_SEEDS", "甜菜种子");
+            put("BEETROOT_SOUP", "甜菜汤");
+            put("BEETROOTS", "甜菜根");
+            put("BELL", "钟");
+            put("BIG_DRIPLEAF", "大型垂滴叶");
+            put("BIG_DRIPLEAF_STEM", "大型垂滴叶茎");
+            put("BIRCH_BOAT", "桦木船");
+            put("BIRCH_BUTTON", "桦木按钮");
+            put("BIRCH_CHEST_BOAT", "桦木箱子船");
+            put("BIRCH_DOOR", "桦木门");
+            put("BIRCH_FENCE", "桦木栅栏");
+            put("BIRCH_FENCE_GATE", "桦木篱笆门");
+            put("BIRCH_PRESSURE_PLATE", "桦木压力板");
+            put("BIRCH_SIGN", "桦树告示牌");
+            put("BIRCH_STAIRS", "桦树楼梯");
+            put("BIRCH_TRAPDOOR", "桦树活板门");
+            put("BIRCH_WALL_SIGN", "桦景墙告示牌");
+            put("BIRCH_WOOD", "白桦木");
+            put("BLACK_CANDLE", "黑色的蜡烛");
+            put("BLACK_CANDLE_CAKE", "黑蜡烛蛋糕");
+            put("BLACK_DYE", "黑色染料");
+            put("BLACK_WALL_BANNER", "黑色墙壁旗帜");
+            put("BLACKSTONE", "黑石");
+            put("BLACKSTONE_SLAB", "黑石台阶");
+            put("BLACKSTONE_STAIRS", "黑石楼梯");
+            put("BLACKSTONE_WALL", "黑石墙");
+            put("BLAST_FURNACE", "高炉");
+            put("BLAZE_SPAWN_EGG", "烈焰人刷怪蛋");
+            put("BLUE_CANDLE", "蓝色的蜡烛");
+            put("BLUE_CANDLE_CAKE", "蓝色的蜡烛的蛋糕");
+            put("BLUE_DYE", "蓝色染料");
+            put("BLUE_WALL_BANNER", "蓝色的墙横幅");
+            put("BRAIN_CORAL_WALL_FAN", "脑珊瑚壁扇");
+            put("BRICK_WALL", "砖墙");
+            put("BRICKS", "砖");
+            put("BROWN_CANDLE", "棕色的蜡烛");
+            put("BROWN_CANDLE_CAKE", "棕色蜡烛的蛋糕");
+            put("BROWN_DYE", "棕色染料");
+            put("BROWN_WALL_BANNER", "棕色的墙横幅");
+            put("BUBBLE_COLUMN", "气泡柱");
+            put("BUBBLE_CORAL_WALL_FAN", "泡沫珊瑚壁扇");
+            put("BUDDING_AMETHYST", "初露头角的紫水晶");
+            put("BUNDLE", "包");
+            put("CALCITE", "方解石");
+            put("CAMPFIRE", "篝火");
+            put("CANDLE", "蜡烛");
+            put("CANDLE_CAKE", "蜡烛的蛋糕");
+            put("CARROT", "胡萝卜");
+            put("CARROT_ON_A_STICK", "胡萝卜钓竿");
+            put("CARROTS", "胡萝卜");
+            put("CARTOGRAPHY_TABLE", "制图学表");
+            put("CAT_SPAWN_EGG", "猫刷怪蛋");
+            put("CAVE_AIR", "洞穴空气");
+            put("CAVE_SPIDER_SPAWN_EGG", "洞穴蜘蛛刷怪蛋");
+            put("CAVE_VINES", "洞穴的葡萄树");
+            put("CAVE_VINES_PLANT", "洞穴藤蔓植物");
+            put("CHAIN", "锁链");
+            put("CHAIN_COMMAND_BLOCK", "命令链块");
+            put("CHARCOAL", "木炭");
+            put("CHEST_MINECART", "运输矿车");
+            put("CHICKEN", "鸡");
+            put("CHICKEN_SPAWN_EGG", "鸡刷怪蛋");
+            put("CHIPPED_ANVIL", "碎铁砧");
+            put("CHISELED_DEEPSLATE", "轮廓分明的深板岩");
+            put("CHISELED_NETHER_BRICKS", "轮廓分明的下面的砖");
+            put("CHISELED_POLISHED_BLACKSTONE", "轮廓分明的抛光黑石");
+            put("CHISELED_STONE_BRICKS", "轮廓分明的石砖");
+            put("CHORUS_FRUIT", "合唱的水果");
+            put("CLOCK", "时钟");
+            put("COBBLED_DEEPSLATE", "深板岩圆石");
+            put("COBBLED_DEEPSLATE_SLAB", "深板岩圆石台阶");
+            put("COBBLED_DEEPSLATE_STAIRS", "深板岩圆石楼梯");
+            put("COBBLED_DEEPSLATE_WALL", "深板岩圆石墙");
+            put("COBBLESTONE_WALL", "圆石墙");
+            put("COBWEB", "蜘蛛网");
+            put("COCOA", "可可");
+            put("COD", "鳕鱼");
+            put("COD_BUCKET", "鳕鱼桶");
+            put("COD_SPAWN_EGG", "鳕鱼刷怪蛋");
+            put("COMMAND_BLOCK", "命令方块");
+            put("COMMAND_BLOCK_MINECART", "命令方块矿车");
+            put("COMPOSTER", "堆肥");
+            put("CONDUIT", "conduit");
+            put("COOKED_COD", "煮熟的鳕鱼");
+            put("COOKED_MUTTON", "煮熟的羊肉");
+            put("COOKED_PORKCHOP", "煮熟的猪排");
+            put("COOKED_RABBIT", "煮熟的兔子");
+            put("COOKED_SALMON", "煮熟的三文鱼");
+            put("COPPER_BLOCK", "铜块");
+            put("COPPER_INGOT", "铜锭");
+            put("COPPER_ORE", "铜矿石");
+            put("CORNFLOWER", "矢车菊");
+            put("COW_SPAWN_EGG", "牛刷怪蛋");
+            put("CRACKED_DEEPSLATE_BRICKS", "裂纹深板岩砖");
+            put("CRACKED_DEEPSLATE_TILES", "裂纹深板岩瓦");
+            put("CRACKED_NETHER_BRICKS", "了下面的砖");
+            put("CRACKED_POLISHED_BLACKSTONE_BRICKS", "抛光裂纹石砖");
+            put("CRACKED_STONE_BRICKS", "裂纹石砖");
+            put("CRAFTING_TABLE", "制作表");
+            put("CREEPER_BANNER_PATTERN", "爬行者横幅图案");
+            put("CREEPER_SPAWN_EGG", "爬行者刷怪蛋");
+            put("CREEPER_WALL_HEAD", "爬行者墙上的头");
+            put("CRIMSON_BUTTON", "绯红木按钮");
+            put("CRIMSON_DOOR", "绯红木门");
+            put("CRIMSON_FENCE", "绯红木栅栏");
+            put("CRIMSON_FENCE_GATE", "绯红木的栅栏门");
+            put("CRIMSON_FUNGUS", "绯红菌");
+            put("CRIMSON_HYPHAE", "绯红菌丝");
+            put("CRIMSON_NYLIUM", "绯红木nylium");
+            put("CRIMSON_PLANKS", "绯红木板");
+            put("CRIMSON_PRESSURE_PLATE", "绯红木压力板");
+            put("CRIMSON_ROOTS", "绯红木的根");
+            put("CRIMSON_SIGN", "绯红木的告示牌");
+            put("CRIMSON_SLAB", "绯红木的台阶");
+            put("CRIMSON_STAIRS", "绯红木的楼梯");
+            put("CRIMSON_STEM", "绯红木的茎");
+            put("CRIMSON_TRAPDOOR", "绯红木的活板门");
+            put("CRIMSON_WALL_SIGN", "绯红木的墙告示牌");
+            put("CROSSBOW", "弩");
+            put("CRYING_OBSIDIAN", "哭泣的黑曜石");
+            put("CUT_COPPER", "切制铜");
+            put("CUT_COPPER_SLAB", "切制铜台阶");
+            put("CUT_COPPER_STAIRS", "切制铜楼梯");
+            put("CUT_RED_SANDSTONE_SLAB", "切制红色砂岩台阶");
+            put("CUT_SANDSTONE_SLAB", "砂岩台阶");
+            put("CYAN_CANDLE", "青色的蜡烛");
+            put("CYAN_CANDLE_CAKE", "青色蜡烛的蛋糕");
+            put("CYAN_WALL_BANNER", "青色的墙横幅");
+            put("DAMAGED_ANVIL", "受损的铁砧");
+            put("DARK_OAK_BOAT", "黑橡树船");
+            put("DARK_OAK_BUTTON", "黑橡树按钮");
+            put("DARK_OAK_CHEST_BOAT", "黑橡树胸部船");
+            put("DARK_OAK_DOOR", "深色的橡木门");
+            put("DARK_OAK_FENCE", "黑橡树栅栏");
+            put("DARK_OAK_FENCE_GATE", "黑橡树篱笆门");
+            put("DARK_OAK_PRESSURE_PLATE", "黑橡树压盘");
+            put("DARK_OAK_SIGN", "黑橡树告示牌");
+            put("DARK_OAK_TRAPDOOR", "黑橡树活板门");
+            put("DARK_OAK_WALL_SIGN", "黑橡树墙告示牌");
+            put("DARK_OAK_WOOD", "黑橡树木材");
+            put("DARK_PRISMARINE_SLAB", "黑暗prismarine台阶");
+            put("DEAD_BRAIN_CORAL_WALL_FAN", "死脑珊瑚墙粉丝");
+            put("DEAD_BUBBLE_CORAL_WALL_FAN", "珊瑚壁扇死泡沫");
+            put("DEAD_FIRE_CORAL_WALL_FAN", "死火珊瑚壁扇");
+            put("DEAD_HORN_CORAL_WALL_FAN", "死角珊瑚壁扇");
+            put("DEAD_TUBE_CORAL_WALL_FAN", "死管珊瑚壁扇");
+            put("DEBUG_STICK", "调试棒");
+            put("DEEPSLATE", "深板岩");
+            put("DEEPSLATE_BRICK_SLAB", "深板岩砖台阶");
+            put("DEEPSLATE_BRICK_STAIRS", "深板岩砖楼梯");
+            put("DEEPSLATE_BRICK_WALL", "深板岩砖墙");
+            put("DEEPSLATE_BRICKS", "深板岩砖");
+            put("DEEPSLATE_COAL_ORE", "深层煤炭矿石");
+            put("DEEPSLATE_COPPER_ORE", "深层铜矿石");
+            put("DEEPSLATE_DIAMOND_ORE", "深层钻石矿");
+            put("DEEPSLATE_EMERALD_ORE", "深层绿宝石矿石");
+            put("DEEPSLATE_GOLD_ORE", "深层金矿");
+            put("DEEPSLATE_IRON_ORE", "深层铁矿石");
+            put("DEEPSLATE_LAPIS_ORE", "深层青金石矿石");
+            put("DEEPSLATE_REDSTONE_ORE", "深层红石矿石");
+            put("DEEPSLATE_TILE_SLAB", "深板岩瓦台阶");
+            put("DEEPSLATE_TILE_STAIRS", "深板岩瓦楼梯");
+            put("DEEPSLATE_TILE_WALL", "深板岩瓦墙");
+            put("DEEPSLATE_TILES", "深板岩瓦砖");
+            put("DIAMOND_HORSE_ARMOR", "钻石马盔甲");
+            put("DIAMOND_SHOVEL", "钻石铲");
+            put("DIORITE_SLAB", "闪长岩台阶");
+            put("DIORITE_STAIRS", "闪长岩楼梯");
+            put("DIORITE_WALL", "闪长岩墙");
+            put("DIRT_PATH", "泥土路径");
+            put("DISC_FRAGMENT_5", "盘碎片5");
+            put("DOLPHIN_SPAWN_EGG", "海豚刷怪蛋");
+            put("DONKEY_SPAWN_EGG", "驴刷怪蛋");
+            put("DRAGON_WALL_HEAD", "龙墙头部");
+            put("DRIED_KELP_BLOCK", "干海带块");
+            put("DRIPSTONE_BLOCK", "石笋块");
+            put("DROPPER", "使滴下的东西");
+            put("DROWNED_SPAWN_EGG", "淹死刷怪蛋");
+            put("ECHO_SHARD", "回声碎片");
+            put("ELDER_GUARDIAN_SPAWN_EGG", "老人守护刷怪蛋");
+            put("ENCHANTED_GOLDEN_APPLE", "附魔金苹果");
+            put("ENCHANTING_TABLE", "附魔台");
+            put("END_GATEWAY", "末地传送门");
+            put("END_PORTAL", "末地传送门");
+            put("END_PORTAL_FRAME", "末地传送门框架");
+            put("END_ROD", "末地烛");
+            put("END_STONE_BRICK_SLAB", "末地砖台阶");
+            put("END_STONE_BRICK_STAIRS", "末地砖楼梯");
+            put("END_STONE_BRICK_WALL", "末地砖墙");
+            put("END_STONE_BRICKS", "末地砖");
+            put("ENDER_EYE", "末影之眼");
+            put("ENDERMAN_SPAWN_EGG", "末影人刷怪蛋");
+            put("ENDERMITE_SPAWN_EGG", "末影螨刷怪蛋");
+            put("EVOKER_SPAWN_EGG", "唤起者刷怪蛋");
+            put("EXPERIENCE_BOTTLE", "经验瓶");
+            put("EXPOSED_COPPER", "暴露的铜");
+            put("EXPOSED_CUT_COPPER", "暴露切制铜");
+            put("EXPOSED_CUT_COPPER_SLAB", "暴露切制铜的台阶");
+            put("EXPOSED_CUT_COPPER_STAIRS", "暴露切制铜楼梯");
+            put("FARMLAND", "农田");
+            put("FILLED_MAP", "填满的地图");
+            put("FIRE", "火");
+            put("FIRE_CHARGE", "火灾负责");
+            put("FIRE_CORAL_WALL_FAN", "火珊瑚壁扇");
+            put("FIREWORK_ROCKET", "烟花火箭");
+            put("FIREWORK_STAR", "烟火之心");
+            put("FLETCHING_TABLE", "制箭台");
+            put("FLOWER_BANNER_PATTERN", "花旗帜图案");
+            put("FLOWER_POT", "花盆");
+            put("FLOWERING_AZALEA", "盛开的杜鹃花");
+            put("FLOWERING_AZALEA_LEAVES", "杜鹃花的叶子开花");
+            put("FOX_SPAWN_EGG", "狐狸刷怪蛋");
+            put("FROG_SPAWN_EGG", "青蛙刷怪蛋");
+            put("FROGSPAWN", "蛙卵");
+            put("FROSTED_ICE", "结霜的冰");
+            put("FURNACE_MINECART", "动力矿车");
+            put("GHAST_SPAWN_EGG", "恶魂刷怪蛋");
+            put("GILDED_BLACKSTONE", "镀金黑石");
+            put("GLASS_PANE", "玻璃板");
+            put("GLISTERING_MELON_SLICE", "闪烁的西瓜片");
+            put("GLOBE_BANNER_PATTERN", "地球横幅图案");
+            put("GLOW_BERRIES", "荧光浆果");
+            put("GLOW_INK_SAC", "荧光墨囊");
+            put("GLOW_ITEM_FRAME", "荧光物品展示框");
+            put("GLOW_LICHEN", "发光地衣");
+            put("GLOW_SQUID_SPAWN_EGG", "荧光乌贼刷怪蛋");
+            put("GOAT_HORN", "山羊的角");
+            put("GOAT_SPAWN_EGG", "山羊刷怪蛋");
+            put("GOLDEN_AXE", "金斧头");
+            put("GOLDEN_BOOTS", "金靴子");
+            put("GOLDEN_CARROT", "金胡萝卜");
+            put("GOLDEN_CHESTPLATE", "黄胸甲");
+            put("GOLDEN_HELMET", "金头盔");
+            put("GOLDEN_HOE", "金锄头");
+            put("GOLDEN_HORSE_ARMOR", "金马护甲");
+            put("GOLDEN_LEGGINGS", "金护腿");
+            put("GOLDEN_PICKAXE", "金镐");
+            put("GOLDEN_SHOVEL", "金铲");
+            put("GOLDEN_SWORD", "金剑");
+            put("GRANITE_SLAB", "花岗石台阶");
+            put("GRANITE_STAIRS", "花岗石楼梯");
+            put("GRANITE_WALL", "花岗岩的墙");
+            put("GRAY_CANDLE", "灰色的蜡烛");
+            put("GRAY_CANDLE_CAKE", "灰色蜡烛的蛋糕");
+            put("GRAY_WALL_BANNER", "灰色的墙壁标语");
+            put("GREEN_CANDLE", "绿色蜡烛");
+            put("GREEN_CANDLE_CAKE", "绿色蜡烛的蛋糕");
+            put("GREEN_WALL_BANNER", "绿墙横幅");
+            put("GRINDSTONE", "磨石");
+            put("GUARDIAN_SPAWN_EGG", "监护刷怪蛋");
+            put("GUNPOWDER", "火药");
+            put("HANGING_ROOTS", "挂根");
+            put("HEAVY_WEIGHTED_PRESSURE_PLATE", "重力压力板");
+            put("HOGLIN_SPAWN_EGG", "hoglin刷怪蛋");
+            put("HONEY_BLOCK", "蜂蜜块");
+            put("HONEY_BOTTLE", "蜂蜜的瓶子");
+            put("HONEYCOMB", "蜂窝");
+            put("HONEYCOMB_BLOCK", "蜂窝块");
+            put("HOPPER", "漏斗");
+            put("HOPPER_MINECART", "漏斗矿车");
+            put("HORN_CORAL_WALL_FAN", "角珊瑚壁扇");
+            put("HORSE_SPAWN_EGG", "马刷怪蛋");
+            put("HUSK_SPAWN_EGG", "皮刷怪蛋");
+            put("INFESTED_CHISELED_STONE_BRICKS", "INFESTED的凿石砖");
+            put("INFESTED_COBBLESTONE", "INFESTED的鹅卵石");
+            put("INFESTED_CRACKED_STONE_BRICKS", "INFESTED了石砖");
+            put("INFESTED_DEEPSLATE", "INFESTED深板岩");
+            put("INFESTED_MOSSY_STONE_BRICKS", "INFESTED长满青苔的石砖");
+            put("INFESTED_STONE", "INFESTED石头");
+            put("INFESTED_STONE_BRICKS", "INFESTED石砖");
+            put("IRON_BARS", "铁棒");
+            put("IRON_HORSE_ARMOR", "铁的马盔甲");
+            put("IRON_NUGGET", "铁金块");
+            put("IRON_SHOVEL", "铁铲子");
+            put("IRON_TRAPDOOR", "铁活板门");
+            put("ITEM_FRAME", "物品框");
+            put("JIGSAW", "拼图");
+            put("JUNGLE_BOAT", "丛林木船");
+            put("JUNGLE_BUTTON", "丛林木按钮");
+            put("JUNGLE_CHEST_BOAT", "丛林木箱子船");
+            put("JUNGLE_DOOR", "丛林木门");
+            put("JUNGLE_FENCE", "丛林木栅栏");
+            put("JUNGLE_FENCE_GATE", "丛林木篱笆门");
+            put("JUNGLE_PRESSURE_PLATE", "丛林木压力板");
+            put("JUNGLE_SIGN", "丛林木告示牌");
+            put("JUNGLE_STAIRS", "丛林木楼梯");
+            put("JUNGLE_TRAPDOOR", "丛林木活板门");
+            put("JUNGLE_WALL_SIGN", "丛林木墙告示牌");
+            put("JUNGLE_WOOD", "丛林木");
+            put("KELP", "海带");
+            put("KELP_PLANT", "海藻植物");
+            put("KNOWLEDGE_BOOK", "知识书");
+            put("LANTERN", "灯笼");
+            put("LARGE_AMETHYST_BUD", "大的紫晶芽");
+            put("LARGE_FERN", "巨大的蕨类植物");
+            put("LAVA", "岩浆");
+            put("LAVA_CAULDRON", "岩浆锅");
+            put("LEAD", "引领");
+            put("LEATHER_HORSE_ARMOR", "马皮护甲");
+            put("LECTERN", "讲台");
+            put("LIGHT", "光");
+            put("LIGHT_BLUE_CANDLE", "浅蓝色的蜡烛");
+            put("LIGHT_BLUE_CANDLE_CAKE", "浅蓝色的蜡烛的蛋糕");
+            put("LIGHT_BLUE_WALL_BANNER", "浅蓝色的墙横幅");
+            put("LIGHT_GRAY_BANNER", "浅灰色的横幅");
+            put("LIGHT_GRAY_BED", "浅灰色的床");
+            put("LIGHT_GRAY_CANDLE", "浅灰色的蜡烛");
+            put("LIGHT_GRAY_CANDLE_CAKE", "浅灰色蜡烛的蛋糕");
+            put("LIGHT_GRAY_CARPET", "浅灰色的地毯");
+            put("LIGHT_GRAY_CONCRETE", "浅灰色混凝土");
+            put("LIGHT_GRAY_CONCRETE_POWDER", "浅灰色混凝土粉");
+            put("LIGHT_GRAY_GLAZED_TERRACOTTA", "浅灰色釉面陶瓦");
+            put("LIGHT_GRAY_SHULKER_BOX", "浅灰色shulker盒子");
+            put("LIGHT_GRAY_STAINED_GLASS", "浅灰色彩色玻璃");
+            put("LIGHT_GRAY_STAINED_GLASS_PANE", "浅灰色彩色玻璃窗格");
+            put("LIGHT_GRAY_WALL_BANNER", "浅灰色墙横幅");
+            put("LIGHT_WEIGHTED_PRESSURE_PLATE", "轻压盘");
+            put("LIGHTNING_ROD", "避雷针");
+            put("LILAC", "淡紫色");
+            put("LILY_OF_THE_VALLEY", "铃兰");
+            put("LILY_PAD", "睡莲叶子");
+            put("LIME_CANDLE", "石灰的蜡烛");
+            put("LIME_CANDLE_CAKE", "石灰蜡烛的蛋糕");
+            put("LIME_WALL_BANNER", "石灰墙上的横幅");
+            put("LLAMA_SPAWN_EGG", "骆驼刷怪蛋");
+            put("LODESTONE", "磁石");
+            put("LOOM", "织机");
+            put("MAGENTA_CANDLE", "洋红色的蜡烛");
+            put("MAGENTA_CANDLE_CAKE", "洋红色的蜡烛蛋糕");
+            put("MAGENTA_WALL_BANNER", "洋红色的墙横幅");
+            put("MAGENTA_WOOL", "洋红色的羊毛");
+            put("MAGMA_CUBE_SPAWN_EGG", "岩浆立方体刷怪蛋");
+            put("MANGROVE_BOAT", "红树林的船");
+            put("MANGROVE_BUTTON", "红树林按钮");
+            put("MANGROVE_CHEST_BOAT", "红树林胸部船");
+            put("MANGROVE_DOOR", "红树林的门");
+            put("MANGROVE_FENCE", "红树林栅栏");
+            put("MANGROVE_FENCE_GATE", "红树林篱笆门");
+            put("MANGROVE_LEAVES", "红树林的叶子");
+            put("MANGROVE_LOG", "红树林的日志");
+            put("MANGROVE_PLANKS", "红树林木板");
+            put("MANGROVE_PRESSURE_PLATE", "红树林压盘");
+            put("MANGROVE_PROPAGULE", "红树林繁殖体");
+            put("MANGROVE_ROOTS", "红树林的根");
+            put("MANGROVE_SIGN", "红树林告示牌");
+            put("MANGROVE_SLAB", "红树林台阶");
+            put("MANGROVE_STAIRS", "红树林楼梯");
+            put("MANGROVE_TRAPDOOR", "红树林活板门");
+            put("MANGROVE_WALL_SIGN", "红树林墙告示牌");
+            put("MANGROVE_WOOD", "红树林木");
+            put("MEDIUM_AMETHYST_BUD", "媒介紫晶芽");
+            put("MELON_SLICE", "瓜切片");
+            put("MELON_STEM", "瓜茎");
+            put("MOJANG_BANNER_PATTERN", "mojang横幅模式");
+            put("MOOSHROOM_SPAWN_EGG", "哞菇刷怪蛋");
+            put("MOSS_BLOCK", "苔藓块");
+            put("MOSS_CARPET", "苔藓地毯");
+            put("MOSSY_COBBLESTONE_SLAB", "长满青苔的鹅卵石台阶");
+            put("MOSSY_COBBLESTONE_STAIRS", "长满青苔的鹅卵石楼梯");
+            put("MOSSY_COBBLESTONE_WALL", "长满青苔的鹅卵石墙");
+            put("MOSSY_STONE_BRICK_SLAB", "长满青苔的石砖台阶");
+            put("MOSSY_STONE_BRICK_STAIRS", "长满青苔的石砖楼梯");
+            put("MOSSY_STONE_BRICK_WALL", "长满青苔的石砖墙");
+            put("MOSSY_STONE_BRICKS", "长满青苔的石砖");
+            put("MOVING_PISTON", "移动的活塞");
+            put("MUD", "泥");
+            put("MUD_BRICK_SLAB", "泥砖台阶");
+            put("MUD_BRICK_STAIRS", "泥砖楼梯");
+            put("MUD_BRICK_WALL", "泥砖墙");
+            put("MUD_BRICKS", "泥砖");
+            put("MUDDY_MANGROVE_ROOTS", "泥泞的红树林根");
+            put("MULE_SPAWN_EGG", "骡子刷怪蛋");
+            put("MUSHROOM_STEW", "蘑菇炖");
+            put("MUTTON", "羊肉");
+            put("NAME_TAG", "姓名标签");
+            put("NETHER_BRICK_FENCE", "下界砖围墙");
+            put("NETHER_BRICK_SLAB", "下界砖台阶");
+            put("NETHER_BRICK_WALL", "下界砖墙");
+            put("NETHER_BRICKS", "下界砖");
+            put("NETHER_GOLD_ORE", "下界金矿石");
+            put("NETHER_PORTAL", "下界传送门");
+            put("NETHER_SPROUTS", "下面的豆芽");
+            put("NETHER_STAR", "下界之星");
+            put("NETHER_WART", "地狱疣");
+            put("NETHERITE_AXE", "下界合金斧");
+            put("NETHERITE_BLOCK", "下界合金块");
+            put("NETHERITE_BOOTS", "下界合金靴子");
+            put("NETHERITE_CHESTPLATE", "下界合金胸甲");
+            put("NETHERITE_HELMET", "下界合金头盔");
+            put("NETHERITE_HOE", "下界合金锄头");
+            put("NETHERITE_INGOT", "下界合金锭");
+            put("NETHERITE_LEGGINGS", "下界合金护腿");
+            put("NETHERITE_PICKAXE", "下界合金镐");
+            put("NETHERITE_SCRAP", "下界合金碎片");
+            put("NETHERITE_SHOVEL", "下界合金铲");
+            put("NETHERITE_SWORD", "下界合金剑");
+            put("OAK_BOAT", "橡树的船");
+            put("OAK_BUTTON", "橡树按钮");
+            put("OAK_CHEST_BOAT", "橡木箱船");
+            put("OAK_DOOR", "橡木门");
+            put("OAK_FENCE", "橡树栅栏");
+            put("OAK_FENCE_GATE", "橡树篱笆门");
+            put("OAK_PRESSURE_PLATE", "橡树压力板");
+            put("OAK_SIGN", "橡树告示牌");
+            put("OAK_TRAPDOOR", "橡树活板门");
+            put("OAK_WALL_SIGN", "橡树墙告示牌");
+            put("OAK_WOOD", "橡木");
+            put("OCELOT_SPAWN_EGG", "豹猫刷怪蛋");
+            put("OCHRE_FROGLIGHT", "赭石霜灯");
+            put("ORANGE_CANDLE", "橙色的蜡烛");
+            put("ORANGE_CANDLE_CAKE", "橙色蜡烛的蛋糕");
+            put("ORANGE_WALL_BANNER", "橙色墙横幅");
+            put("OXIDIZED_COPPER", "氧化铜矿");
+            put("OXIDIZED_CUT_COPPER", "氧化切制铜");
+            put("OXIDIZED_CUT_COPPER_SLAB", "氧化切制铜的台阶");
+            put("OXIDIZED_CUT_COPPER_STAIRS", "氧化切制铜楼梯");
+            put("PACKED_MUD", "挤泥");
+            put("PANDA_SPAWN_EGG", "熊猫刷怪蛋");
+            put("PARROT_SPAWN_EGG", "鹦鹉刷怪蛋");
+            put("PEARLESCENT_FROGLIGHT", "珠母froglight");
+            put("PEONY", "牡丹");
+            put("PETRIFIED_OAK_SLAB", "石化橡木台阶");
+            put("PHANTOM_SPAWN_EGG", "幻影刷怪蛋");
+            put("PIG_SPAWN_EGG", "猪刷怪蛋");
+            put("PIGLIN_BANNER_PATTERN", "猪灵横幅图案");
+            put("PIGLIN_BRUTE_SPAWN_EGG", "猪灵刷怪蛋");
+            put("PIGLIN_SPAWN_EGG", "猪灵刷怪蛋");
+            put("PILLAGER_SPAWN_EGG", "掠夺者刷怪蛋");
+            put("PINK_CANDLE", "粉色的蜡烛");
+            put("PINK_CANDLE_CAKE", "粉色的蜡烛蛋糕");
+            put("PINK_WALL_BANNER", "粉红墙上标语");
+            put("PISTON_HEAD", "活塞头");
+            put("PLAYER_WALL_HEAD", "玩家墙的头");
+            put("PODZOL", "灰壤");
+            put("POINTED_DRIPSTONE", "尖石笋");
+            put("POISONOUS_POTATO", "毒土豆");
+            put("POLAR_BEAR_SPAWN_EGG", "北极熊刷怪蛋");
+            put("POLISHED_ANDESITE_SLAB", "抛光安山岩台阶");
+            put("POLISHED_ANDESITE_STAIRS", "抛光安山岩楼梯");
+            put("POLISHED_BASALT", "抛光玄武岩");
+            put("POLISHED_BLACKSTONE", "抛光黑石");
+            put("POLISHED_BLACKSTONE_BRICK_SLAB", "抛光黑石砖台阶");
+            put("POLISHED_BLACKSTONE_BRICK_STAIRS", "抛光黑石砖楼梯");
+            put("POLISHED_BLACKSTONE_BRICK_WALL", "抛光黑石砖墙");
+            put("POLISHED_BLACKSTONE_BRICKS", "抛光黑石砖");
+            put("POLISHED_BLACKSTONE_BUTTON", "抛光黑石按钮");
+            put("POLISHED_BLACKSTONE_PRESSURE_PLATE", "抛光黑石压盘");
+            put("POLISHED_BLACKSTONE_SLAB", "抛光黑石台阶");
+            put("POLISHED_BLACKSTONE_STAIRS", "抛光黑石楼梯");
+            put("POLISHED_BLACKSTONE_WALL", "抛光黑石墙");
+            put("POLISHED_DEEPSLATE", "抛光深板岩");
+            put("POLISHED_DEEPSLATE_SLAB", "抛光深板岩台阶");
+            put("POLISHED_DEEPSLATE_STAIRS", "抛光深板岩楼梯");
+            put("POLISHED_DEEPSLATE_WALL", "抛光深板岩墙");
+            put("POLISHED_DIORITE_SLAB", "抛光闪长岩台阶");
+            put("POLISHED_DIORITE_STAIRS", "楼梯的闪长岩");
+            put("POLISHED_GRANITE_SLAB", "抛光花岗岩台阶");
+            put("POLISHED_GRANITE_STAIRS", "抛光花岗岩台阶");
+            put("POPPED_CHORUS_FRUIT", "突然合唱水果");
+            put("PORKCHOP", "猪排");
+            put("POTATO", "土豆");
+            put("POTATOES", "土豆");
+            put("POTTED_ACACIA_SAPLING", "盆栽合金欢树苗");
+            put("POTTED_ALLIUM", "盆栽葱属植物");
+            put("POTTED_AZALEA_BUSH", "盆栽杜鹃花丛");
+            put("POTTED_AZURE_BLUET", "盆栽天蓝色矢车菊");
+            put("POTTED_BAMBOO", "盆栽竹");
+            put("POTTED_BIRCH_SAPLING", "盆栽白桦树苗");
+            put("POTTED_BLUE_ORCHID", "盆栽蓝兰花");
+            put("POTTED_BROWN_MUSHROOM", "盆栽棕色蘑菇");
+            put("POTTED_CACTUS", "盆栽的仙人掌");
+            put("POTTED_CORNFLOWER", "盆栽的矢车菊");
+            put("POTTED_CRIMSON_FUNGUS", "盆栽深红色的真菌");
+            put("POTTED_CRIMSON_ROOTS", "盆栽深红色的根");
+            put("POTTED_DANDELION", "盆栽蒲公英");
+            put("POTTED_DARK_OAK_SAPLING", "盆栽黑橡树树苗");
+            put("POTTED_DEAD_BUSH", "布什盆栽死了");
+            put("POTTED_FERN", "盆栽蕨类植物");
+            put("POTTED_FLOWERING_AZALEA_BUSH", "盆栽花卉的杜鹃花");
+            put("POTTED_JUNGLE_SAPLING", "盆栽丛林树苗");
+            put("POTTED_LILY_OF_THE_VALLEY", "盆栽莉莉的山谷");
+            put("POTTED_MANGROVE_PROPAGULE", "盆栽红树林繁殖体");
+            put("POTTED_OAK_SAPLING", "盆栽橡树树苗");
+            put("POTTED_ORANGE_TULIP", "盆栽橙色郁金香");
+            put("POTTED_OXEYE_DAISY", "盆栽大眼睛戴西");
+            put("POTTED_PINK_TULIP", "盆栽粉色郁金香");
+            put("POTTED_POPPY", "盆栽罂粟");
+            put("POTTED_RED_MUSHROOM", "盆栽红蘑菇");
+            put("POTTED_RED_TULIP", "盆栽红郁金香");
+            put("POTTED_SPRUCE_SAPLING", "盆栽云杉树苗");
+            put("POTTED_WARPED_FUNGUS", "盆栽扭曲的真菌");
+            put("POTTED_WARPED_ROOTS", "盆栽扭曲的根源");
+            put("POTTED_WHITE_TULIP", "盆栽白色郁金香");
+            put("POTTED_WITHER_ROSE", "盆栽玫瑰枯萎");
+            put("POWDER_SNOW", "粉雪");
+            put("POWDER_SNOW_BUCKET", "粉雪斗");
+            put("POWDER_SNOW_CAULDRON", "粉雪大锅");
+            put("PRISMARINE_BRICK_SLAB", "海晶石砖台阶");
+            put("PRISMARINE_BRICK_STAIRS", "海晶石砖楼梯");
+            put("PRISMARINE_CRYSTALS", "海晶石晶体");
+            put("PRISMARINE_SHARD", "海晶石碎片");
+            put("PRISMARINE_SLAB", "海晶石台阶");
+            put("PRISMARINE_WALL", "海晶石墙");
+            put("PUFFERFISH", "河豚");
+            put("PUFFERFISH_BUCKET", "河豚桶");
+            put("PUFFERFISH_SPAWN_EGG", "河豚刷怪蛋");
+            put("PUMPKIN_PIE", "南瓜饼");
+            put("PUMPKIN_STEM", "南瓜干");
+            put("PURPLE_CANDLE", "紫色的蜡烛");
+            put("PURPLE_CANDLE_CAKE", "紫色蜡烛的蛋糕");
+            put("PURPLE_WALL_BANNER", "紫色的墙壁标语");
+            put("PURPUR_SLAB", "紫台阶");
+            put("QUARTZ", "石英");
+            put("QUARTZ_BLOCK", "石英块");
+            put("QUARTZ_BRICKS", "石英砖");
+            put("QUARTZ_SLAB", "石英台阶");
+            put("RABBIT", "兔子");
+            put("RABBIT_HIDE", "兔子藏");
+            put("RABBIT_SPAWN_EGG", "兔刷怪蛋");
+            put("RABBIT_STEW", "兔子炖");
+            put("RAIL", "铁路");
+            put("RAVAGER_SPAWN_EGG", "破坏者刷怪蛋");
+            put("RAW_COPPER", "生铜");
+            put("RAW_COPPER_BLOCK", "生的铜块");
+            put("RAW_GOLD", "生金");
+            put("RAW_GOLD_BLOCK", "生黄金块");
+            put("RAW_IRON", "生的铁");
+            put("RAW_IRON_BLOCK", "生铁块");
+            put("RECOVERY_COMPASS", "复苏的指南针");
+            put("RED_CANDLE", "红色的蜡烛");
+            put("RED_CANDLE_CAKE", "红蜡烛的蛋糕");
+            put("RED_DYE", "红色染料");
+            put("RED_NETHER_BRICK_SLAB", "红色的砖台阶");
+            put("RED_NETHER_BRICK_STAIRS", "砖红色的楼梯");
+            put("RED_NETHER_BRICK_WALL", "红色的砖墙");
+            put("RED_SANDSTONE_SLAB", "红色砂岩台阶");
+            put("RED_SANDSTONE_WALL", "红色砂岩墙");
+            put("RED_WALL_BANNER", "红色的墙横幅");
+            put("REDSTONE_BLOCK", "雷石东块");
+            put("REDSTONE_LAMP", "雷石东灯");
+            put("REDSTONE_WALL_TORCH", "雷石东墙火炬");
+            put("REDSTONE_WIRE", "雷石东线");
+            put("REINFORCED_DEEPSLATE", "坚固深板岩");
+            put("REPEATING_COMMAND_BLOCK", "重复命令块");
+            put("RESPAWN_ANCHOR", "重生锚");
+            put("ROOTED_DIRT", "扎根泥土");
+            put("ROSE_BUSH", "玫瑰丛中");
+            put("SALMON", "大马哈鱼");
+            put("SALMON_BUCKET", "大马哈鱼桶");
+            put("SALMON_SPAWN_EGG", "鲑鱼刷怪蛋");
+            put("SANDSTONE_SLAB", "砂岩台阶");
+            put("SANDSTONE_WALL", "砂岩墙");
+            put("SCAFFOLDING", "脚手架");
+            put("SCULK", "沉船");
+            put("SCULK_CATALYST", "sculk催化剂");
+            put("SCULK_SENSOR", "sculk传感器");
+            put("SCULK_SHRIEKER", "sculk尖叫");
+            put("SCULK_VEIN", "sculk静脉");
+            put("SCUTE", "鳞甲");
+            put("SHEEP_SPAWN_EGG", "绵羊刷怪蛋");
+            put("SHROOMLIGHT", "蘑菇灯");
+            put("SHULKER_SHELL", "shulker壳");
+            put("SHULKER_SPAWN_EGG", "shulker刷怪蛋");
+            put("SILVERFISH_SPAWN_EGG", "银色的鱼刷怪蛋");
+            put("SKELETON_HORSE_SPAWN_EGG", "骷髅马刷怪蛋");
+            put("SKELETON_SPAWN_EGG", "骨架刷怪蛋");
+            put("SKELETON_WALL_SKULL", "头骨骷髅墙");
+            put("SKULL_BANNER_PATTERN", "头骨横幅模式");
+            put("SLIME_BLOCK", "粘液阻塞");
+            put("SLIME_SPAWN_EGG", "粘液刷怪蛋");
+            put("SMALL_AMETHYST_BUD", "小紫晶芽");
+            put("SMALL_DRIPLEAF", "小垂滴叶");
+            put("SMITHING_TABLE", "锻造台");
+            put("SMOKER", "烟");
+            put("SMOOTH_BASALT", "光滑的玄武岩");
+            put("SMOOTH_QUARTZ_SLAB", "光滑的石英台阶");
+            put("SMOOTH_QUARTZ_STAIRS", "光滑的石英楼梯");
+            put("SMOOTH_RED_SANDSTONE_SLAB", "光滑的红色砂岩台阶");
+            put("SMOOTH_RED_SANDSTONE_STAIRS", "光滑的红色砂岩楼梯");
+            put("SMOOTH_SANDSTONE_SLAB", "光滑的砂岩台阶");
+            put("SMOOTH_SANDSTONE_STAIRS", "光滑的砂岩楼梯");
+            put("SMOOTH_STONE_SLAB", "光滑的石台阶");
+            put("SNOW", "雪");
+            put("SNOWBALL", "雪球");
+            put("SOUL_CAMPFIRE", "灵魂篝火");
+            put("SOUL_FIRE", "灵魂火");
+            put("SOUL_LANTERN", "灵魂灯笼");
+            put("SOUL_SOIL", "灵魂的土壤");
+            put("SOUL_TORCH", "灵魂火炬");
+            put("SOUL_WALL_TORCH", "灵魂墙火炬");
+            put("SPAWNER", "已成熟的雌鱼");
+            put("SPIDER_SPAWN_EGG", "蜘蛛刷怪蛋");
+            put("SPORE_BLOSSOM", "孢子开花");
+            put("SPRUCE_BOAT", "云杉的船");
+            put("SPRUCE_BUTTON", "云杉按钮");
+            put("SPRUCE_CHEST_BOAT", "云杉胸部船");
+            put("SPRUCE_DOOR", "云杉门");
+            put("SPRUCE_FENCE", "云杉栅栏");
+            put("SPRUCE_FENCE_GATE", "云杉篱笆门");
+            put("SPRUCE_PRESSURE_PLATE", "云杉压盘");
+            put("SPRUCE_SIGN", "云杉告示牌");
+            put("SPRUCE_STAIRS", "云杉楼梯");
+            put("SPRUCE_TRAPDOOR", "云杉活板门");
+            put("SPRUCE_WALL_SIGN", "云杉墙告示牌");
+            put("SPRUCE_WOOD", "云杉木材");
+            put("SPYGLASS", "小望远镜");
+            put("SQUID_SPAWN_EGG", "乌贼刷怪蛋");
+            put("STONE_BRICK_WALL", "石砖墙");
+            put("STONE_BRICKS", "石砖");
+            put("STONE_PRESSURE_PLATE", "石头压盘");
+            put("STONE_SHOVEL", "石头铲");
+            put("STONE_SLAB", "石台阶");
+            put("STONE_STAIRS", "石阶");
+            put("STONECUTTER", "石匠");
+            put("STRAY_SPAWN_EGG", "杂散刷怪蛋");
+            put("STRIDER_SPAWN_EGG", "水黾刷怪蛋");
+            put("STRIPPED_CRIMSON_HYPHAE", "剥夺了深红色的菌丝");
+            put("STRIPPED_CRIMSON_STEM", "剥夺了深红色的茎");
+            put("STRIPPED_MANGROVE_LOG", "剥夺了红树林的日志");
+            put("STRIPPED_MANGROVE_WOOD", "剥夺了红树林木");
+            put("STRIPPED_WARPED_HYPHAE", "被扭曲的菌丝");
+            put("STRIPPED_WARPED_STEM", "被扭曲的茎");
+            put("STRUCTURE_BLOCK", "结构块");
+            put("STRUCTURE_VOID", "结构空隙");
+            put("SUNFLOWER", "向日葵");
+            put("SUSPICIOUS_STEW", "可疑的炖肉");
+            put("SWEET_BERRIES", "甜蜜的浆果");
+            put("SWEET_BERRY_BUSH", "布什甜浆果");
+            put("TADPOLE_BUCKET", "蝌蚪桶");
+            put("TADPOLE_SPAWN_EGG", "蝌蚪刷怪蛋");
+            put("TALL_SEAGRASS", "高大的海草");
+            put("TARGET", "目标");
+            put("TINTED_GLASS", "有色玻璃");
+            put("TNT_MINECART", "TNT矿车");
+            put("TRADER_LLAMA_SPAWN_EGG", "交易员骆驼刷怪蛋");
+            put("TRAPPED_CHEST", "被困的箱子");
+            put("TRIPWIRE", "绊线");
+            put("TROPICAL_FISH", "热带鱼");
+            put("TROPICAL_FISH_BUCKET", "热带鱼桶");
+            put("TROPICAL_FISH_SPAWN_EGG", "热带鱼类刷怪蛋");
+            put("TUBE_CORAL_WALL_FAN", "管珊瑚壁扇");
+            put("TUFF", "凝灰岩");
+            put("TURTLE_EGG", "龟蛋");
+            put("TURTLE_SPAWN_EGG", "龟刷怪蛋");
+            put("TWISTING_VINES", "扭葡萄");
+            put("TWISTING_VINES_PLANT", "缠绕藤蔓植物");
+            put("VERDANT_FROGLIGHT", "翠绿的froglight");
+            put("VEX_SPAWN_EGG", "烦恼刷怪蛋");
+            put("VILLAGER_SPAWN_EGG", "村民刷怪蛋");
+            put("VINDICATOR_SPAWN_EGG", "维护者刷怪蛋");
+            put("VOID_AIR", "无效的空气");
+            put("WALL_TORCH", "墙火炬");
+            put("WANDERING_TRADER_SPAWN_EGG", "流浪的交易员刷怪蛋");
+            put("WARDEN_SPAWN_EGG", "监狱长刷怪蛋");
+            put("WARPED_BUTTON", "扭曲的按钮");
+            put("WARPED_DOOR", "扭曲的门");
+            put("WARPED_FENCE", "扭曲的栅栏");
+            put("WARPED_FENCE_GATE", "扭曲的篱笆门");
+            put("WARPED_FUNGUS", "扭曲的真菌");
+            put("WARPED_FUNGUS_ON_A_STICK", "弯曲的棍子上的真菌");
+            put("WARPED_HYPHAE", "扭曲的菌丝");
+            put("WARPED_NYLIUM", "扭曲nylium");
+            put("WARPED_PLANKS", "扭曲的木板");
+            put("WARPED_PRESSURE_PLATE", "扭曲压盘");
+            put("WARPED_ROOTS", "扭曲的根源");
+            put("WARPED_SIGN", "扭曲告示牌");
+            put("WARPED_SLAB", "扭曲的台阶");
+            put("WARPED_STAIRS", "弯曲的楼梯");
+            put("WARPED_STEM", "扭曲的茎");
+            put("WARPED_TRAPDOOR", "扭曲的活板门");
+            put("WARPED_WALL_SIGN", "扭曲的墙告示牌");
+            put("WARPED_WART_BLOCK", "扭曲的疣块");
+            put("WATER", "水");
+            put("WATER_CAULDRON", "水的大锅");
+            put("WAXED_COPPER_BLOCK", "上过蜡的铜块");
+            put("WAXED_CUT_COPPER", "蜡切制铜");
+            put("WAXED_CUT_COPPER_SLAB", "蜡切制铜的台阶");
+            put("WAXED_CUT_COPPER_STAIRS", "蜡切制铜楼梯");
+            put("WAXED_EXPOSED_COPPER", "蜡暴露铜");
+            put("WAXED_EXPOSED_CUT_COPPER", "蜡暴露切制铜");
+            put("WAXED_EXPOSED_CUT_COPPER_SLAB", "蜡暴露切制铜的台阶");
+            put("WAXED_EXPOSED_CUT_COPPER_STAIRS", "蜡暴露切制铜楼梯");
+            put("WAXED_OXIDIZED_COPPER", "蜡氧化铜矿");
+            put("WAXED_OXIDIZED_CUT_COPPER", "蜡氧化切制铜");
+            put("WAXED_OXIDIZED_CUT_COPPER_SLAB", "蜡氧化切制铜的台阶");
+            put("WAXED_OXIDIZED_CUT_COPPER_STAIRS", "蜡氧化切制铜楼梯");
+            put("WAXED_WEATHERED_COPPER", "蜡风化铜");
+            put("WAXED_WEATHERED_CUT_COPPER", "蜡风化切制铜");
+            put("WAXED_WEATHERED_CUT_COPPER_SLAB", "蜡风化降低铜的台阶");
+            put("WAXED_WEATHERED_CUT_COPPER_STAIRS", "蜡风化切制铜楼梯");
+            put("WEATHERED_COPPER", "饱经风霜的铜");
+            put("WEATHERED_CUT_COPPER", "风化切制铜");
+            put("WEATHERED_CUT_COPPER_SLAB", "风化切制铜的台阶");
+            put("WEATHERED_CUT_COPPER_STAIRS", "风化切制铜楼梯");
+            put("WEEPING_VINES", "哭泣的藤蔓");
+            put("WEEPING_VINES_PLANT", "哭泣的藤蔓植物");
+            put("WHEAT_SEEDS", "小麦种子");
+            put("WHITE_CANDLE", "白色蜡烛");
+            put("WHITE_CANDLE_CAKE", "白蜡烛蛋糕");
+            put("WHITE_DYE", "白色染料");
+            put("WHITE_WALL_BANNER", "白墙横幅");
+            put("WITCH_SPAWN_EGG", "女巫刷怪蛋");
+            put("WITHER_ROSE", "枯萎的玫瑰");
+            put("WITHER_SKELETON_SPAWN_EGG", "枯萎的骨架刷怪蛋");
+            put("WITHER_SKELETON_WALL_SKULL", "枯萎头骨骷髅墙");
+            put("WOLF_SPAWN_EGG", "狼刷怪蛋");
+            put("WOODEN_AXE", "木斧");
+            put("WOODEN_HOE", "木制的锄头");
+            put("WOODEN_PICKAXE", "木制的鹤嘴锄");
+            put("WOODEN_SHOVEL", "木铲");
+            put("WOODEN_SWORD", "木刀");
+            put("WRITABLE_BOOK", "可写的书");
+            put("YELLOW_CANDLE", "黄色的蜡烛");
+            put("YELLOW_CANDLE_CAKE", "黄色的蜡烛的蛋糕");
+            put("YELLOW_WALL_BANNER", "黄色的墙壁标语");
+            put("ZOGLIN_SPAWN_EGG", "猪灵刷怪蛋");
+            put("ZOMBIE_HORSE_SPAWN_EGG", "僵尸马刷怪蛋");
+            put("ZOMBIE_SPAWN_EGG", "僵尸刷怪蛋");
+            put("ZOMBIE_VILLAGER_SPAWN_EGG", "僵尸村民刷怪蛋");
+            put("ZOMBIE_WALL_HEAD", "僵尸墙的头");
+            put("ZOMBIFIED_PIGLIN_SPAWN_EGG", "“僵尸化”猪灵刷怪蛋");
         }
     };
 
