@@ -51,7 +51,7 @@ public class AffairPage implements ViewPage {
         }
         ItemStack item = SerializeUtil.deserialize(sale.getStack(), sale.getMeta());
         // 展示物品
-        itemMap.put(10, item);
+        itemStackMap.put(10, item);
 
         // 下架按钮
         ItemStack cancel;
@@ -67,7 +67,7 @@ public class AffairPage implements ViewPage {
             ItemUtils.addLore(cancel, "§3§l上架于§7:§f "+ TimeUtils.getTime(sale.getAppear()));
             ItemUtils.setItemValue(cancel, "cancel "+sale.getId());
         }
-        itemMap.put(17, cancel);
+        itemStackMap.put(17, cancel);
 
         // 添加交易操作按钮
         if(sale.isAuction()){
@@ -77,21 +77,21 @@ public class AffairPage implements ViewPage {
             ItemUtils.addLore(add1, "§3§l增加§7:§f 10");
             ItemUtils.addLore(add1, "§3§l当前§7:§f "+sale.getCost());
             ItemUtils.setItemValue(add1, "add "+sale.getId()+" 10");
-            itemMap.put(13, add1);
+            itemStackMap.put(13, add1);
 
             ItemStack add2 = MarketItem.ADD_INGOT.clone();
             ItemUtils.setDisplayName(add2, "§e§l加价");
             ItemUtils.addLore(add2, "§3§l增加§7:§f 100");
             ItemUtils.addLore(add2, "§3§l当前§7:§f "+sale.getCost());
             ItemUtils.setItemValue(add2, "add "+sale.getId()+" 100");
-            itemMap.put(14, add2);
+            itemStackMap.put(14, add2);
 
             ItemStack add3 = MarketItem.ADD_BLOCK.clone();
             ItemUtils.setDisplayName(add3, "§e§l加价");
             ItemUtils.addLore(add3, "§3§l增加§7:§f 1000");
             ItemUtils.addLore(add3, "§3§l当前§7:§f "+sale.getCost());
             ItemUtils.setItemValue(add3, "add "+sale.getId()+" 1000");
-            itemMap.put(15, add3);
+            itemStackMap.put(15, add3);
         }else {
             // 直售
             int amount = item.getAmount();
@@ -100,21 +100,21 @@ public class AffairPage implements ViewPage {
             ItemUtils.addLore(buy1, "§3§l价格§7:§f "+ ConvertUtils.formatDouble(sale.getCost()/amount));
             ItemUtils.setItemValue(buy1, "buy "+sale.getId()+" 1");
             buy1.setAmount(1);
-            itemMap.put(13, buy1);
+            itemStackMap.put(13, buy1);
             if(amount > 8){
                 ItemStack buy2 = MarketItem.BUY_8.clone();
                 ItemUtils.setDisplayName(buy2, "§e§l部分");
                 ItemUtils.addLore(buy2, "§3§l价格§7:§f "+ ConvertUtils.formatDouble(sale.getCost()/amount * 8));
                 ItemUtils.setItemValue(buy2, "buy "+sale.getId()+" 8");
                 buy2.setAmount(8);
-                itemMap.put(14, buy2);
+                itemStackMap.put(14, buy2);
             }
             ItemStack buy3 = MarketItem.BUY_ALL.clone();
             ItemUtils.setDisplayName(buy3, "§e§l一口价");
             ItemUtils.addLore(buy3, "§3§l价格§7:§f "+ ConvertUtils.formatDouble(sale.getCost()));
             ItemUtils.setItemValue(buy3, "buy "+sale.getId());
             buy3.setAmount(amount);
-            itemMap.put(15, buy3);
+            itemStackMap.put(15, buy3);
         }
         return itemStackMap;
     }
